requirements.txt
# 跨平台自动化项目依赖包

# 跨平台GUI自动化 (替代pywin32)
pyautogui>=0.9.54
pynput>=1.7.6

# 图像处理和计算机视觉
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# 系统进程管理
psutil>=5.9.0

# 跨平台窗口管理
pygetwindow>=0.0.9

# 多进程和并发
# 标准库已包含 threading, multiprocessing, concurrent.futures

# 日志记录
# 标准库已包含 logging

# 数据结构和类型提示
# 标准库已包含 dataclasses, typing, enum

# 时间处理
# 标准库已包含 time, datetime

# UUID生成
# 标准库已包含 uuid

# 队列操作
# 标准库已包含 queue

# 抽象基类
# 标准库已包含 abc

# 系统路径操作
# 标准库已包含 os, sys

# 开发和测试工具（可选）
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# 文档生成（可选）
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
